server-app/requirements.txt
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.0.7
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
exceptiongroup==1.2.2
h11==0.14.0
httptools==0.6.1
idna==3.7
kombu==5.4.0
msgpack==1.0.8
oauthlib==3.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
vine==5.1.0
watchfiles==0.23.0
wcwidth==0.2.13
websockets==12.0
